backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
alembic==1.12.1
python-multipart==0.0.6
cryptography==41.0.7
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1